Fix missing omd libssl dependency when executing python during build
Several steps during the build procedure execute the python command. All
of them use the python we build for omd, but fail to execute it with an
environment where the required libopenssl is available.

Some of these call sites even six from python3-modules to be available.

This commit centralizes the command computation and also centralizes the
make target dependencies. All you need to do now is to use
$(PACKAGE_PYTHON_COMMAND) while setting $(PACKAGE_PYTHON_COMMAND_DEPS
dependency.

Change-Id: I542104c5dfb10b27f77b96724d55111840cb2952