13401 mk_mongodb.py adapt logging
In previous versions the mk_mongodb.py agent plugin printed expected
exceptions to stderr. This may confuse the user as they may think
something is wrong with the plugin or the mongodb server while this
exception is totally fine.

For example mk_mongodb.py queries for replica information. Is a mongodb
server queried that has no replica configured, the plugin wrote an error
message to stderr that it could not fetch the replication Information.

These Errors are now hidden by default. They may be shown by specifying
`-vv` when executing the agent plugin manually.

The errors are logged with log level debug.

CMK-8658

Change-Id: I657f414e4b0250e07e79e3f0ee5d5f31d3bcdbc3