13284 FIX BI: Reintroduced "sitename#hostname" filter for "Regex for hostname
You may add the prefix <tt>{sitename}#</tt> to any <tt>Regex for hostname</tt> filter, e.g
<tt>central_site#server_(.*)"<tt>
This will further restrict the filter so that it only includes hosts for the given site.

This feature is more or less undocumented and has been reintroduced to fix errors with older configurations.
We do not recommend to use this special prefix and plan to offer a better alternative in a future version.

Change-Id: I6af34774fd0af22c42b02f7abb157e0594c695f3