13363 Veeam: Add CDP job monitoring
You can now monitor the status and the time since last execution of the
Continuous Data Protection (CDP) jobs run by Veeam.

In order to carry out the check, the Veeam agent plugin
needs to be installed on the backup server.

For further information look at the respective manpage

SUP-7811

Change-Id: Ib1d679c931507ab05f47b621ada1cab0aa90a83e