13355 FIX OpenWRT agent: Make <tt>df</tt> section work with BusyBox
BusyBox does not support <tt>-x</tt> and <tt>-l</tt> as arguments to the
<tt>df</tt> command. These arguments are removed from the OpenWRT agent,
s.t. the <tt>df</tt> section is now available in combination with BusyBox.

Furthermore, <tt>-i</tt> is sometimes unsupported as well, so the agent
will now first check if this argument is supported and only attempt to
collect inodes information if it is.

Note that in order to avoid hanging agents due to remote filesystems,
the <tt>df</tt> calls are terminated after a fixed timeout of 5 s. This
means that the agent will only attempt to produce the <tt>df</tt> section
if the <tt>waitmax</tt> command is available. Checkmk provides a
<tt>waitmax</tt> binary or alternatively, you can install the
<tt>timeout</tt> command.

Closes #410

Change-Id: Ic0a787271f01c820c92be554248d5972e8fa76a6