tcp fetcher: centralize transport validation
This commit is mostly a refactoring, bundling up the logging
and the validation of the detected transport protocol,
respectively.

If we're using TLS, we accept both encrypted or unencrypted
data. This is done to ease migration from 2.0 to 2.1.
If deployed via agent bakery the agent will not additionaly
encrypt the data below the TLS layer.

Change-Id: I33ccb28c41cc03c2d4c612ba6974a92764cba8d8