Extract Visual Filter class
For text, regex, dropdown & ajax dropdown

TODO: Dare for the moment not to validate the regex, because user can only
select what comes back. In general we should validate all input although that
is not yet implemented. Only the regex had a control, and the dropdown display
was a hack

Change-Id: I510e7770d0e7e6996dfd6fb697662505ffa2cc7d