Improve build caching of 3to2ed agent plugins
During creation of the source packages we update the Checkmk version
using "make setversion" from the base directory. This updates the
Checkmk version of the agent .py files and also the mtime of the .py
files which results in "make" rebuild all _2.py files even if they
didn't change compared to the last time.

Since we want to optimize the time of the source package build, we
introduce this to use the source file hash to decide whether or not to
update the _2.py files. Not ideal, but works. Let me know if you have
something

Change-Id: I9e41461be2af57f433bbf038684aa56d4f4ce858