simplify failed decryption case
'decrypt_by_agent_protocol' will not throw an MKFetcherError.
If we did not detect the transport protocol to be 'PLAIN' but neither
succeeded in decrypting the output, we raise unconditionally.
We no longer assume that we might have misinterpreted plaintext data
as being encrypted (because of missing '<<<' at the beginning).

Change-Id: I24154bde82d92db8b5bc544d24b650ed13dfdf90