FilterGroupCombo is trully only an AjaxDropdown
With this change all the group filters use the ajax autocompleter and
the are close to equivalent to how grafana behaves.

request_vars_from_row is parasitic as values are not really gathered
from rows. Yet, I still don't know where it came into use case to
isolate it.

Heading info should use the same autocompleter with the extra livestatus
filtering. Still pending how to pass the operator for this query type.

Because the options are resolved by ajax, no livestatus query happens on
render, thus also drop some of those queries on test.

Change-Id: Icb7269a5061bc8675e2c549b75c6375dec781f51