13453 apc_symmetra_elphase: Remove default current threshold
apc_symmetra_elphase check would assume CRIT state for current values bigger then 1 A.
Since this value depends on the configuration of the UPS and the battery type, it
produced error states for normal current values.

If you want to keep using 1 A as a threshold value for apc_symmetra_elphase current
set it up in the check's parameters.

SUP-8517

Change-Id: Ic55303b998a59c5c7cb22a10276dc070c463590c