Packaging: Do not compile test core and livestatus test files
Previously the packaging of Checkmk was calling "make all" of
enterprise/core/src and livestatus/src which was, besides the required
binaries, also compiling test_neb and test_cmc binaries together with
googletest.

This change nearly halves the compilation times of both and may even
have more effect with an increasing number of tests.

Change-Id: I0dd86193b30524b31f1df794879b892dd9f36f44