LDAP: Search for contactgroup aliases as well as names
Since contactgroups names have a very limited character set, it was
impossible to add users to an LDAP group which had e.g. a space in its
CN. By also searching LDAP for contactgroup aliases, which have broader
character set, this becomes possible!

To be backwards compatible, contactgroup names are preferred: If there
are two contactgroups, one with name "x" and the other with alias "x",
the one with name "x" will be the one users are added to if they are in
an LDAP group with CN "x".