Make cmk.gui.plugins an implicit namespace package
To be able to extend Checkmk with GUI plugin modules using the local
hierarchy, we need to make the different plugin packages implicit
namespace packages (according to PEP 402).

This is exact the same mechanism we use for our agent based plugins.
We "only" need to move some stuff out of the way to make it possible
to remove the __init__.py files from the different plugin packages.

This will be done step by step with the upcoming commits.

Change-Id: I75d69b1e9997ccb5591583b850592bbbe01e340b