Make cmk.gui.plugins.dashboard an implicit namespace package
To be able to extend Checkmk with GUI plugin modules using the local
hierarchy, we need to make the different plugin packages implicit
namespace packages (according to PEP 402).

This is the first package that has been converted to the new mechanism.
With this change it is now possible to put local dashboard plugins into
the path <tt>lib/check_mk/cmk/gui/plugins/dashboard</tt>.

Dashboard plugins were previously loaded from
<tt>local/share/check_mk/web/plugins/dashboard</tt>. This will still
work for 2.1, but will be removed one day.

Change-Id: I3c1fbdeb088b24bfd19622d92db28cf82656aa7c