Fix walking of nested packages
Walking nested packages did not work as expected because
we missed to provide the `prefix` argument.

The doc string of `pkgutil.walk_packages` is actually a bit misleading.
Because it says "prefix is a string to output" one could understand it
to be only relevant for the data structures reported.

But when  `pkgutil.walk_packages` comes across packages, it also uses
the `prefix` in it's internal imports.

Once setting `prefix`, the recursion works as expected.

Change-Id: Idb6c09497f021caf30515477058d4f5099dd401a