13130 FIX mk_oracle: Fix bakery crash on deactivated agent rule
When configuring one or more "ORACLE databases" agent rules and deactivating
it with by setting "Activation" to "Do not deploy ORACLE databases plugin",
the agent bakery failed to bake agents.

This is fixed now by moving the activation/deactivation to an entry of the
ruleset. As a result, for every active agent configuration, at least one rule
with specified "Activation" entry must exist to successfully bake agents.

CMK-9080

Change-Id: I3586eb728b2d531ae3a1baaeba7cf7e6fa640286