13569 Make ntopng integration vlan aware
In the past, the ntopng integration could not deal with vlan aware ntopng instances.
A Host in ntopng is in this case suffixed with its vlan id, e.g.: 10.1.2.3@4
With this werk, the ntopng integration into checkmk is now vlan aware.
Keep in mind, that this feature is still experimental as we're trying to gain more information about such setups.

Change-Id: Ib334a562ab97fbd9924f92f0146ea1785383182d