13321 SEC NagVis: Updated to 1.9.29 (Fix possible deletion of arbitrary files)
Fix possible deletion of arbitrary files (CVE-2021-33178).

An authenticated user with enough permissions to access the NagVis.
ManageBackgrounds endpoint, such as admin, can delete arbitrary files on the
server limited by the rights of the Apache system user. In Checkmk, this is
limited to files owned by the site user.

https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2021-33178

Change-Id: I4a88322a91e8b4ba86f662a09133d068c548e586