13476 FIX Occasional KeyError in inline SNMP fetcher
A race condition could lead to an occasional KeyError reported by the inline
SNMP fetcher.

The re-assigning of an empty dict in the cleanup routine would in fact
only clear the global vaiable *in that module* (bound to its local name
during import), not the variable from the utils module.

The actual bug has already been fixed during a "refactoring", but we
make the offending global variable private anyway.

Change-Id: I1e8320f928576b265858b8af2890008999e36f2b