Add CA of a site to trusted CAs and verify CA chain in encrypted mknotifyd connections
When a site is created the CA of the site is added to the "Trusted
certificate authorities for SSL" in the global settings. If a site is
updated the CA of the site is added if it is not a remote site with
replication. The site CA can then be used to verify encrypted SSL
connections of the mknotifyd without further user interaction.

CMK-9131

Change-Id: Iafc10b8492f8fd115cd7bf79688f120bd8985ce8