13323 FIX Fix validation of host addresses in different situations
There are different situations where host addresses can be configured within
the Checkmk setup (e.g. Additional IP addresses, host diagnostic, management
address). In some of them host names, IPv4 and IPv6 addresses are allowed and
in some of them only a subset. Depending on the situation, the validation
should be as specific as possible, but the inserted value was not correctly
validated in all the situations.

Change-Id: If472c46a93d94118ede42e560eeea37f0c3dd3d5