Fix IP address missmatch for datasources
Commit d3193de86a9dc8efd376e5e343140789c3be30ea started to
use the sources IP address for the source key in
`cmk.base.agent_base.data_provider._collect_host_sections`,
because it seemed they should and would always match.

They did not. In case of a host without IP address, the
datasource would be provided with a fallback IP ("0.0.0.0"
or "::0"). This resulted in missing sections for all
related plugins.

Since the datasource is the client, there is no point in
distinguishing between `None` and zero, so we pass on
`None` as it is less surprising.

Note that this variable is passed to the plugins that
provide the commandline call of the agent. Had we ever had
a proper specification of the API, this would be an
incompatible change. As a courtesy, we provide a werk.

Let's hope this dosen't break too much.

Change-Id: Idc067dd8b4d5a482d1a09a0362b7a0d97c93d510