13324 SEC Shipping software bill of materials with Checkmk
A software bill of materials can be used to analyze the components that are
part of Checkmk. An accurate inventory of all components enables organizations
to identify risk, allows for greater transparency, and enables rapid impact
analysis.

Checkmk packages are now shipping a BOM in OWASP CycloneDX format. As of
Checkmk 2.1.0 and 2.0.0p18 it can be found in each site at the path
<tt>share/doc/bill-of-materials.json</tt>.

We also publish the BOM to internal dependency track instance.

Change-Id: I6dc73b7accc5eaf237fa9c69828bb7bf11092563