Allow custom caching prefixes in local checks
THIS IT MAY CHANGE AT ANY TIME

Cached local checks can now include their custom prefix in
the `cached(${CACHED_AT},${VALID_FOR})` format.
This allows for various workarounds for problems caused by
the fact that the validitiy period and the data generation
interval are always the same, which leads to periodically
occurring moments without valid data.

This is only an intermediate solution, until we find a
proper way of addressing this. You must not rely on this
behaviour, which is why there won't be a werk for this.

Change-Id: I4df7a242137364de53ddf7272ecb9f7dc5eed927