Migrate check_command filter
Resolved the service_check_command collision. Issue was that the "ident"
of selected tag is also taken as the "div" container class. If on top of
that "check_command" is the css for finding the select2 element and the
ajax endpoint target, it will happen that select2 edits the "div" tag
instead of the "select" tag and damage the DOM.

Change-Id: I7b2cf08ba99a2e772d77fa9f30ddb025c2790b84