memory_linux: Reuse UpperMemoryLevels for Kubernetes
In order to reuse th valuespec UpperMemoryLevels for Kubernetes
it's necessary to able to configure default values, so they
match the default values used within the check.
We reuse the valuespec in order to achieve a uniform look across
the new kube_memory check and the Memory and Swap usage on Linux
rule.

Change-Id: Ice5e6047fcec5a87b657d0913406e74d205e892d