Split Checkbox filters
Render component is a row of checkboxes
- The "_hidden" varprefix is not needed when all request_vars are
collected. The livestatus filter funtions didn't use this anymore, thus
remove it from the render logic.

Filter Logic is a Multioption filter. Each case has an independent
function filter and the available options are in the query_filters
modules. That will make them later available in the webapi.

TODO: Check if I can make the options_toggle_filter a default. Maybe a
new subclass

Change-Id: I78e21550a604dbde85c6a74e4ea8bf0df5aa8881