Add test to ensure the Pipfile.lock is in sync with Pipfile
We now had an outdated Pipfile.lock multiple times which then
resulted in the version build jobs recognizing this issue with
the message:

```
Pipfile.lock (f9092b) out of date, updating to (cdb517)...
```

Which then failed because the build job is not using the `--pre` flag
which we currently use for building our Pipfile.lock because black is
only releasing preview releases. Anyway. Even if that would work, then
we don't want to create a new lock during the version builds.

Now, to detect this as early as possible, there is this test.

Change-Id: I85e2732ae8799b90025dce26aac934d381c7b7a6