13404 FIX mk_docker.py: diskstat: do not discovery devices without counters
Before this change the discovery behaviour of the diskstat check was
dependent on the source of the data: When running an agent inside a
docker container, devices without counters where not discovered, when
running the mk_docker.py agent plugin those devices where discovered.

With this Werk devices without counters are no longer discovered with
mk_docker.py

The main purpose of this commit was to add tests before moving
docker_container_diskstat to the new check api.

Change-Id: I1278d1e72126af343da3312964895787d3cf1fc6