Sites Filters
They are rather special as they don't work at the livestatus query
level, but as an external options. For that reason don't use the
AjaxDropdown with TextFilter, but a separate instance.

On the same idea, MultipleSites is a separate renderer instead of
inherinting from dual list.

Finally, make a query for sites an ajax endpoint.

Currently all this endpoints are growning out of control. Need to reduce
the amount of endpoints as the way they are activate on valuespecs.js

Change-Id: Iee78e752fb82614214bff7fe0063cfcef8c47e3d