13091 FIX Fixed event console time filters
The event console contains logic to quickly skip history files which are not
relevant for a given query. This logic was broken from day one, often
resulting in superfluous processing of irrelevant history files, which in
turn could lead to e.g. timeouts in the GUI. This has been fixed now.

Change-Id: I25d105e2daf4aa76c424ca4cff54f4a5bf792cf0