Improve when mk_zypper is run
mk_zypper takes a long time to run, so most people will set it up to run asynchronously every X hour. The problem with this is the alert stays in cache until the plugin is executed again, which can take a long time. Instead of this, it would be better to detect whether there has been any update on the system (with /var/log/zypp/history) and to then run the check again. If there has been no change, we still want to check at least once a day if there is anything new. There are probably better ways to implement this (I'm using a cache file and a spool file) but this is working fine on OpenSUSE Leap 15.3.