Use Query as class name to separate filters from renderers
The GUI previously only had Filters. Which were a large class that does
the logical filtering plus the HTML rendering. After the recent
separation it is clear that an isolated Query class can existis taking
the responisbility of the logical filtering.

The Other part being the renderer is not yet renamed as the Filter
architecture is not completely removed.

Change-Id: I37b8f094745cb1786f65e8c952ed018437fcb091