shellcheck fix SC2169
I'm not sure if all of the 'dash' findings are
relevant for ash. Just fix them anyway.

SC2169: In dash, &> is not supported.
SC2169: In dash, echo flags besides -n not supported.
SC2169: In dash, HOSTNAME is not supported.
SC2169: In dash, EUID is not supported.
SC2169: In dash, /dev/{tcp,udp} is not supported.

Change-Id: I3a3b5e9f28c31fca999735d45c17b4964293c15e