Add backup codes for two-factor authentication
In case a user does not have the authentication device at
hand, the backup codes can be used as an alternative.

The codes can be generated on the user profile page and are
only shown once to the user. During generation they are stored
in the user profile in the same form as regular passwords.

In the moment a backup code is used, it is automatically invalidated.

Change-Id: Ib20e87f7ab2b74d9b58888818af786ca399979ff