Use only the Logfile part of the dereferenced iterators.
The LogCache's map contains time_point/Logfile pairs, where the time_point
is always the Logfile's since(), so there is no need for the first part of
the iterator data.

This is a tiny cleanup step towards a simpler iteration: We effectively need
only a range of Logfiles. If we only had ranges... :-/

Change-Id: I62d827dde12650de889abc071d70a2393dcf339b