Cleanup composition, integration, GUI crawl tests dependency on node
The test execution does not require node, npm, node_modules. The reason
why this dependency was there, that there is the option to update the
Checkmk version under test with ".f12". In the moment "web/.f12" is
executed, these things are required. But only then.

For this reason, we leave it to the web/.f12 script to care about
everything in the moment it is executed. And in fact it is already
doing this.

This change should allow us to remove nodejs from most of our CI
containers. We only need in the ubuntu-20.04 container because it
executes the 'source package build' and tests where we need nodejs.

Change-Id: Idc1613c8fcce0639f992c4b5abb4f074b941903f