13197 FIX Add space as valid URL character
With Werk 13195 a stricter URL validation was introduced. In the original
implementation the space character was not allowed. Unfortunately servicenames
with spaces are not encoded properly and spaces used to work. Since this is
also common behaviour and does not impose security risks the space is allowed
again.

It seems that we are quite lax with URL encoding our links and browsers
are very tolerant. The RFC also asks for robustness for user-typed URIs.
Therefore allowing the space character as valid URI character.

Change-Id: I1044e7cd84aa3acf145cbdd493cdd3873393c2c7