test_kube_pod_restarts: Optionally return an empty fixture
The test in question takes advantage of a mutable fixture (a dictionary
in this case) to assert one aspect of correctness. Basically, it tests
one expected side-effect of the check function.

Relying on a fixture and overriding its object makes resilience tests
fail, as running it once leaves the mutable object in an "invalid"
state for the next iteration.

What makes it an issue is the way with which `pytest-repeat` "tricks"
tests into repeating themselves by adding one additional dimension of
parametrization, which does not trigger a re-assignment of the other
parameters and a consequential reset to their initial states.

To overcome that, the fixture is changed to rely on a flag that tells
whether it should return an empty dict or not. The test can then
override this flag instead of the fixture itself.

Change-Id: I89224cc587dca1ca4d390a3ad32172b99f42eeb8