Remove binary config code from Python
Rough remove of the code to create to binary config.

Two additional changes
- get_strategy() is moved to microcore as a best solution to
fix typisation error: _sequential_strategy returns removed
type
- pre-fixed problem with indexes, this change will disappear
on rebase and done to avoid flaky testing in gerrit

This is the first commit in the raw. It is planned to move all
functionality from core_cmc to microcore_config.

TEST: manual in heute with cmk -O cmk-U ans so on

CMK-9168

Change-Id: Ia44d6632214e26f3ca4b45edd0745e62d6f608bc