13198 Stop showing automation secrets
Before this Werk, on the edit_user page the automation secret could be viewed.
This could be abused by other vulnerabilities (e.g. XSS) to retrieve this
secret to abuse it later.

When creating a new automation user / edit an automation secret you should
write the secret down (e.g. in a password store).

Change-Id: I727394d94669078e9021b3fd700f399a74af523c