Improve GUI crawl speed
The combined_graphs.py page renders graphs. In case the page is opened
with an unrestricted context (no filter or lax filters), the page displays
a large number of graphs. These pages take more than 10 seconds, some
even 30 seconds to load which is way too long for the crawler.

We now skip all these pages and only call combined_graphs.py in case we
find a host context filter in the URL.

Change-Id: Ic4da3d322acc42a2a689e13477729b362b3aa8c6