13082 FIX scheduling host downtimes with REST API from read-only sites
This Werk fixes a bug which was introduced in Werk 13080.

When a site was used as "read-only", (or any other setup which allows
only access to Livestatus but doesn't have a config) setting downtimes
was not possible.

This Werk fixes this. Setting downtimes is now also possible:

* on sites with no configuration
* for hosts which have already been deleted in the configuration, but the change has not been activated yet.

No user intervention is necessary to make this work.

This Werk refers to commit ccb385fd5278b277d8ca3775ef5a0adc368eb453

Change-Id: I030162b9db9e3c769994ebb243ed0f58f8f14d51