13405 FIX Allow to disable thread levels
The rule "Number of threads" can set both relative and absolute levels
for the number of threads. Up until this Werk it was not possible to
remove said levels completely once they where set.

Now this is possible. For example you can define a base rule matching
all hosts to check for absolute levels and add another rule matching a
single host removing the absolute levels from the base rule and only
check for relative levels. To accomplish that choose "No levels" for the
absolute levels on the single host.

SUP-9106

Change-Id: I4ef4c893b6a85498fbf881c0003907b6c5df3e5f