13632 FIX Password store: Fix sharing passwords with other contact group members
In the password store it is possible to save a password entry and share it with members
of other contact groups. The members of these groups can then use the passwords in their
Checkmk configuration.

The sharing based on contact group memberships did not work at all, because the
matching of the contact groups a user has did not work. The result is that a
shared password would not be displayed for members of the contact group.

Change-Id: I775a01be63736bcb7348fc31d120e57e684abe66