Refactoring: Disentangle TableStateHistory and TableCachedStatehist, part 2.
Split off the code for adding statehist columns, we will need it at another
place later.

Note that we are a little bit more general than necessary (no need for a
prefix), but this is being done for consistency reasons with other
addColumns() functions. Perhaps we can formalize this later.

Change-Id: Ib9381f34f962c06c3ba90ad0188edee6efae71ba