Remove passwords from passwords.mk
The Password data structure we save in passwords.mk now holds the meta
data attributes related to a password store entry. To be consistent with
our Password type we leave the 'password' attribute with an empty string
value in the passwords.mk.

The actual passwords are stored exclusively in the stored_passwords
file.

With this change the stored_passwords file is directly updated when
saving a password change in Setup instead of during the core
configuration creation. To get back the previous behaviour we'll change
the base and core code to read the passwords from a separate file which
gets only updated during the core configuration creation in the next
step.

Change-Id: I9a27f76a31cde56955892daf7a189cec2ce6791c