Fix graph titles of single metric graphs
The 2.1 introduced this regression. Adding the title here prevented the
graph from using the correct title. The title is computed later during
processing of the graph. At this point of the graph processing it is not
easily possible to compute the graph title.

The type GraphTemplate treats the title title as optional attribute
fro other situations and in pre 2.1 code it was done exactly like
we do it now.

Change-Id: I3b92adbc3ac98fe6869124d199e855f5f502d602