13717 SEC Persistant XSS in Predefined Conditions
This Werk fixes a Persistant Cross-Site-Scripting (XSS) vulnerability. (CWE-79)

The title of a <i>Predefined condition</i> is not properly escaped when shown
as condition.

No mitigation is available.

Checkmk 1.6 and Checkmk 2.0 were subject to this vulnerability.

To detect if this vulnerability is/was used you can check
<tt>etc/check_mk/conf.d/wato/predefined_conditions.mk</tt> for HTML code.
Please be aware that an attacker could delete the code after a attack.

CVE is requested and will be added later.

CVSS: CVSS:3.1/AV:N/AC:L/PR:L/UI:R/S:U/C:L/I:H/A:N (6.3 medium)

We thank Manuel Sommer for finding this vulnerability and bringing this to our
attention.

Change-Id: Id48483af0639af06ea901e9916877b752da80b70