13406 FIX agent_vsphere: KeyError: runtime.host
Before this werk the special agent `agent_vsphere` crashed when
it was executed while a virtual machine is cloned. The freshly cloned
virtual machine is not assigned to any vmware host, and thus can not be
reporetd.

If a virtual machine is reported to have no vmware host, the virtual
machine is now ignored and a message is printed to stderr.

SUP-9157

Change-Id: I3b460e4c58a6039341d60e19b370e61d654ab430