13614 FIX missing context in title on hostgroup and servicegroup views
On the pursuit to avoid querying the URL which may include the linked
view request vars on last minute and only select from context it is not
possible to find filters from a linked view without a datasource
available. Because link filters only happen when switching
from (host/service)group datasource to host/service datasource. As this
function is datasource unaware we optionally test for this posibility
when (host/service)group is a single info.

SUP-8637

Change-Id: I207d0d4875be2be8fec2bab9368c91acc4f1f3c0