Drop get_all_package_infos() helper
This function was useful some time ago for wrapping all the needed
information in a single data object for the GUI when we used automation
calls for transporting the information from base to the GUI code.

Since the GUI can now directly import cmk.utils.packaging, the GUI can
directly access the relevant functions.

Add some more type hints for the GUI part and fixing the unhandled case
of broken or missing MKP packages in the GUI.

Change-Id: Ic8883170c6bb4295200ef4bfdb040f0e1ac26a44