MKP installation and removal trigger a GUI plugin reload again
Previous versions of the GUI automatically monitored their plugins for
changes and tried to reload them in case a change was detected. This was
costly and error prone and implemented with some kind of own plugin
loading logic. The mechanism has now been replaced with our standard
plugin loader.

But we still have to apply changes to the plugins automatically.
Otherwise the user would have to restart / reload the GUI web server
after each MKP installation, which is not acceptable for a pure GUI
user.

We now ensure that the GUI is automatically loading new plugins by
triggering a reload of the site apache in case an MKP contains GUI
related files.

Change-Id: I07e15a94af3294a92d00bf947a19a16abebf77d7