13634 FIX REST API: Creating hosts via API does not trigger agent baking by default anymore
The REST API endpoints for creating hosts (Create a host, Create a cluster
host, Bulk create hosts) were triggering and agent baking job for the new
hosts after configuring the hosts.

This may be useful in some cases, but in the most cases it is not intended to
do it. Since the job needs some resources for the baking and also holds locks
during baking, the job might slow down subsequent API calls.

To address both needs we now added an optional query parameter
<tt>?bake_agent=[1|0]</tt> to the API calls mentioned above.

The default behaviour has been changed: If you don't specify the parameter, the
agents are not baked anymore. Only when you call the endpoints with
<tt>?bake_agent=1</tt>, the agent baking job will be started.

CMK-9553

Change-Id: I461296a27abebaac26aafcb799d7437c18444562