Sneak CMK 2.1 autocompleter endpoints for Grafana
The rewrited Grafana connector makes use the VisualContext filtering and
many of the new autocompleter endpoints.

The VisualContext filtering is treated by directly stashing the context
received as a parameter from Grafana into the URI vars, as it is needed
on 2.0. Since the context comes directly from Grafana there is no
conflict of sources and no need to backport the entire Filter refactor
that happend on 2.1. Still, the filtering procedures as "sneaked" into
the codebase in isolation instead of integrating the regular filter logic.

Then some autocompletion endpoints are added for Grafana to have the
data avalible. On existing autocomplete monitored_metrics the
VisualContext option is "sneaked" in. For monitored_host &
monitored_service_description, since they are defined in
cmk.gui.valuespec importing from a plugin module would create a
nightmare on import cycles and are thus not modified. Users would have
data, but won't enjoy the filtering of it of all other active filters.
They would get as many suggestions as they normally get on CMK 2.0,
where filters are independent on the FilterMenu

Change-Id: I7af8090a7cb6e743aaf8d4d96f5b72e95a24b7ae