Integration tests: Remove changing core during CEE tests
We now have dedicated CRE integration test jobs which cares about testing
the CRE independently of the CEE jobs.

For this reason we completely remove the core switching during the
tests. This will give us a huge performance improvement of the CEE jobs.
The total execution time is nearly halfed now.

We obviously don't test the Nagios core in CEE sites anymore then, but
this is not a relevant use case anyways, so we can live with that gap.

Change-Id: I7378ac476c9aa902bfadd5ffe03823f4d0cf9c57